pyproject.toml
[project]
name = "orchestrator"
version = "0.2.0"
description = "Generic orchestrator for running Claude skills and agents across multiple repositories"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "fastmcp>=2.13.1",
    "python-dotenv>=1.0.0",
]

[project.scripts]
orchestrator = "orchestrator.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
packages = ["src/orchestrator"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
